Extract hdfs_glob and reuse it in orc_write_test
